+++ src/lib.rs
@@ -227,3 +227,4 @@ pub mod datalog;
 pub mod error;
 pub mod format;
 pub mod token;
+pub mod wasm;
